created a new ProductPage to render it as a client component
